requirements.txt
discord.py
aiohttp
python-dotenv
aiomysql